[Test] Made async exec tests obey async rule.
Previously the execution tests for the async calling convention ignored
the rule that callers of async functions must themselves be async.
Failing to obey that rule prevented landing of the SIL verification
change that enforces that rule.  Here, the tests are modified to satisfy
that rule so that the SIL verification change can be landed.