[sil-combine] Update for OSSA.
I was able to update everything except for a few small transformations around
casts that it is not clear how to optimize them in OSSA without some new
instructions. I don't think they are important optimizations, but if they are,
we can find a way to optimize them if we need to.