Use Relative Path to Test Inputs
Unfortunately, OutputFileMap does not normalize input names when it is constructing keys. Instead, keys to the map are formed by appending the entry's value directly onto the working directory with the test host's path separator convention.

On Windows, the tests here were using UNIX-style path separators when passing inputs to the Driver, so the key was

"C:\Path\To\File\A.swift"

But the input path to the driver was

"C:\Path\To\File/A.swift"

To work around this - and to ultimately make the test slightly more portable - just use relative paths to the input files since we've already changed directories to the working directory before we run these driver commands.