Sema: Use VarDecl::isLayoutExposedToClients() when checking access in @frozen structs
We require that all stored properties in a @frozen struct have
public or @usableFromInline types, even if the property itself
is not public. This is so that clients can correctly generate
code to manipulate the @frozen struct.

This check was only looking for bona-fide stored properties,
and missing out looking at properties that have backing storage,
namely 'lazy' properties and property wrappers.