[IRGen] Track repeated requirements for async functions in NecessaryBindings.
Previously, NecessaryBindings used a SetVector to store the
GenericRequirements that it accumulates.  That was a problem for the
layout of async contexts where it is possible for the same generic
argument to be repeated.  Meanwhile, that is correct for partial
application forwarders where there is a single thunk for every partial
apply and consequently the bindings can be packed in without
duplication.

Here, a SetVector is used only when the NecessaryBindings are for
partial apply forwarders.  When the NecessaryBindings are instead for
async functions, a SmallVector is used.