DI: Fix crash when value_metatype was used with a non-class 'self' type
In a constructor, SILGen lowers type(of: self) to a value_metatype
instruction, but 'self' might not have been initialized yet if
this is a convenience initializer, so we replace it with a use of
the 'self' metatype argument.

However getDynamicSelfMetadata() asserts if 'self' is a non-class
type. Since we know we're inside of a proper method here and not a
closure that captured 'self', we can use getSelfArgument() instead.

Fixes https://bugs.swift.org/browse/SR-12665 / rdar://problem/62481587.