SIL: Look through SILBoxType when checking for invalid non-escaping function value captures
While you can't write a noescape function type explicitly except
on a function-typed parameter, you can still assign the value of
such a parameter to another 'let' or 'var' with an inferred type.

We correctly handled the 'let' case, but in the 'var' case the
value is wrapped inside a SILBoxType, so we have to look through
that.

Fixes <rdar://problem/70822136>.