[android] Remove XFAIL for tests that started passing.
These tests were disabled in #33500 (a0333a4e370) and #30390
(6ac04c6079) because the optimizer did not seem to be working in
Android ARMv7.

At some moment between October 21st and October 30th something unbroke
the tests. I have looked into the changes in the SILOptimizer directory,
but I cannot pinpoint it to any of them in particular. It might be also
changes introduced in LLVM.