[android] Split demangling tests for Android 32/64 bits.
`size_t` differs in 32 bit platforms like Android ARMv7 where it is
`unsigned int` instead of `unsigned long`. The mangling changes one of the
characters.

The original tests are disabled in Android, and they are replicated in
different files for Android, with both the 32 and the 64 bits versions
of the mangling and the function signatures.

The original tests are not modified in order to avoid complicated checks
to avoid platforms like iphonesimulator-i386, which is a 32 bit
platform, but still uses `unsigned long` as the underlying type for
`size_t`

Both tests started failing after #34057 landed in the Android 32 bits CI
machines.