Re-add #30710
Reverts the reversion in #34188.

Bisection points to #34188 as the cause of a crasher that looks to be the
result of a miscompile. Adding this back resolves the crasher.

Details of the crash:
The `DiscontiguousSliceSlicing` test in
validation-test/StdlibUnittest/RangeSet.swift crashes in the `every()`
function on line 15. When the contents of this function are replaced with a
functional equivalent that uses different syntax, the crash resolves.