[ownership] Add a few test cases that show that the reborrow verifier handles recursive reborrows correctly.
I also removed this tests dependency on the Swift module which speeds up running
with a debug compiler.

I am finishing up a generic RAUW value1 with value2 regardless of ownership. I
want to depend on this behavior, so I wanted to put a test into tree to make
sure it works/keeps on working.